requirements.txt
pycryptodome==3.9.7
jsbeautifier
requests==2.23.0
art==4.5
cloudscraper==1.2.33
beautifulsoup4==4.8.2
js2py==0.68
PySimpleGUI==4.18.0
polling==0.3.1
